requirements.txt
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.0
distlib==0.3.9
docstring_parser==0.16
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
firebase-admin==5.2.0 
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.86.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
matplotlib-inline==0.1.7
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13